arreglo

+++ controllers/deleteUserControllers.js
@@ -3,8 +3,14 @@ const router = express.Router();
 const conexion = require('../models/conetion');
 
 router.get('/deleteUser/:cedula', (req, res) => {
+    // Validar sesión
+    if (!req.session.login) {
+        const mensaje = 'Por favor, inicia sesión para acceder a esta página.';
+        return res.redirect('/login?mensaje=' + encodeURIComponent(mensaje));
+    }
+
     let mensaje;
-    const  cedula  = req.params.cedula;
+    const cedula = req.params.cedula;
 
     const sql = `DELETE FROM usuario WHERE cedula = ?`;
     conexion.query(sql, [cedula], (err, result) => {
